Refactor ExpandableListView
