+++ Example/App.js
@@ -13,7 +13,7 @@ import {
   Platform,
   Linking
 } from 'react-native';
-import { Expandable_ListView } from './ExpandableListView';
+import { ExpandableListView } from './ExpandableListView';
 import { Actions } from './constants';
 import * as utils from './utils';
 
@@ -338,7 +338,418 @@ export default class App extends Component {
     ];
   }
 
-  update_Layout = (index) => {
+  onItemPress = (item) => {
+    switch (item.action) {
+      case Actions.SET_USER_PROFILE:
+        utils.set_userProfile();
+        break;
+      case Actions.SET_MULTI_VALUES:
+        CleverTap.profileSetMultiValuesForKey(['a', 'b', 'c'], 'letters');
+        break;
+      case Actions.REMOVE_MULTI_VALUE:
+        CleverTap.profileRemoveMultiValueForKey('b', 'letters');
+        break;
+      case Actions.ADD_MULTI_VALUE:
+        CleverTap.profileAddMultiValueForKey('d', 'letters');
+        break;
+      case Actions.INCREMENT_VALUE:
+        CleverTap.profileIncrementValueForKey(10, 'score');
+        CleverTap.profileIncrementValueForKey(3.141, 'PI_Float');
+        CleverTap.profileIncrementValueForKey(
+          3.141592653589793,
+          'PI_Double',
+        );
+        break;
+      case Actions.DECREMENT_VALUE:
+        CleverTap.profileDecrementValueForKey(10, 'score');
+        CleverTap.profileDecrementValueForKey(3.141, 'PI_Float');
+        CleverTap.profileDecrementValueForKey(
+          3.141592653589793,
+          'PI_Double',
+        );
+        break;
+      case Actions.USER_LOGIN:
+        utils.onUser_Login();
+        break;
+      case Actions.CLEVERTAP_ID:
+        utils.getCleverTap_id();
+        break;
+      case Actions.USER_LOCATION:
+        utils.set_userLocation();
+        break;
+      case Actions.USER_LOCALE:
+        utils.set_Locale();
+        break;
+      case Actions.INITIALIZE_INBOX:
+        CleverTap.initializeInbox();
+        break;
+      case Actions.SHOW_INBOX:
+        utils.show_appInbox();
+        break;
+      case Actions.SHOW_INBOX_TABS:
+        utils.show_appInboxwithTabs();
+        break;
+      case Actions.INBOX_TOTAL_MESSAGE_COUNT:
+        utils.get_TotalMessageCount();
+        break;
+      case Actions.INBOX_UNREAD_MESSAGE_COUNT:
+        utils.get_UnreadMessageCount();
+        break;
+      case Actions.INBOX_ALL_MESSAGES:
+        utils.Get_All_InboxMessages();
+        break;
+      case Actions.INBOX_UNREAD_MESSAGES:
+        utils.get_All_InboxUnreadMessages();
+        break;
+      case Actions.INBOX_MESSAGE_FOR_ID:
+        utils.Get_InboxMessageForId();
+        break;
+      case Actions.INBOX_DELETE_MESSAGE_FOR_ID:
+        utils.delete_InboxMessageForId();
+        break;
+      case Actions.INBOX_READ_MESSAGE_FOR_ID:
+        utils.markRead_InboxMessageForId();
+        break;
+      case Actions.INBOX_NOTIFICATION_VIEWED:
+        utils.pushInboxNotificationViewed();
+        break;
+      case Actions.INBOX_NOTIFICATION_CLICKED:
+        utils.pushInboxNotificationClicked();
+        break;
+      case Actions.PUSH_EVENT:
+        utils.pushevent();
+        break;
+      case Actions.PUSH_CHARGED_EVENT:
+        utils.pushChargedEvent();
+        break;
+      case Actions.SET_DEBUG:
+        CleverTap.setDebugLevel(3);
+        break;
+      case Actions.CREATE_NOTIFICATION_GROUP:
+        utils.create_NotificationChannelGroup();
+        break;
+      case Actions.CREATE_NOTIFICATION_CHANNEL:
+        utils.create_NotificationChannel();
+        break;
+      case Actions.DELETE_NOTIFICATION_CHANNEL:
+        utils.delete_NotificationChannel();
+        break;
+      case Actions.DELETE_NOTIFICATION_GROUP:
+        utils.delete_NotificationChannelGroup();
+        break;
+      case Actions.PUSH_FCM:
+        utils.pushFcmRegistrationId();
+        break;
+      case Actions.CREATE_NOTIFICATION:
+        utils.create_notification();
+        break;
+      case Actions.CREATE_NOTIFICATION_CHANNEL_WITH_SOUND:
+        utils.createNotificationChannelWithSound();
+        break;
+      case Actions.CREATE_NOTIFICATION_CHANNEL_WITH_GROUP:
+        utils.createNotificationChannelWithGroupId();
+        break;
+      case Actions.CREATE_NOTIFICATION_CHANNEL_WITH_GROUP_AND_SOUND:
+        utils.createNotificationChannelWithGroupIdAndSound();
+        break;
+      case Actions.DISPLAY_UNIT_ID:
+        utils.getUnitID();
+        break;
+      case Actions.ALL_DISPLAY_UNITS:
+        utils.getAllDisplayUnits();
+        break;
+      case Actions.PRODUCT_CONFIG_FETCH:
+        utils.fetch();
+        break;
+      case Actions.PRODUCT_CONFIG_ACTIVATE:
+        utils.activate();
+        break;
+      case Actions.PRODUCT_CONFIG_FETCH_AND_ACTIVATE:
+        utils.fetchAndActivate();
+        break;
+      case Actions.PRODUCT_CONFIG_FETCH_INTERVAL:
+        utils.fetchwithMinIntervalinsec();
+        break;
+      case Actions.PRODUCT_CONFIG_SET_INTERVAL:
+        utils.setMinimumFetchIntervalInSeconds();
+        break;
+      case Actions.PRODUCT_CONFIG_GET_BOOL:
+        utils.getBoolean();
+        break;
+      case Actions.PRODUCT_CONFIG_GET_DOUBLE:
+        utils.getDouble();
+        break;
+      case Actions.PRODUCT_CONFIG_GET_LONG:
+        utils.getLong();
+        break;
+      case Actions.PRODUCT_CONFIG_GET_STRING:
+        utils.getString();
+        break;
+      case Actions.PRODUCT_CONFIG_GET_STRINGS:
+        utils.getStrings();
+        break;
+      case Actions.PRODUCT_CONFIG_RESET:
+        utils.reset_config();
+        break;
+      case Actions.PRODUCT_CONFIG_LAST_FETCH_TS:
+        utils.getLastFetchTimeStampInMillis();
+        break;
+      case Actions.GET_FEATURE_FLAG:
+        utils.getFeatureFlag();
+        break;
+      case Actions.IN_APPS_SUSPEND:
+        CleverTap.suspendInAppNotifications();
+        break;
+      case Actions.IN_APPS_DISCARD:
+        CleverTap.discardInAppNotifications();
+        break;
+      case Actions.IN_APPS_RESUME:
+        CleverTap.resumeInAppNotifications();
+        break;
+      case Actions.ENABLE_PERSONALIZATION:
+        utils.enablePersonalization();
+        break;
+      case Actions.ATTRIBUTION_IDENTIFIER:
+        utils.GetCleverTapAttributionIdentifier();
+        break;
+      case Actions.OPT_OUT:
+        CleverTap.setOptOut(false);
+        break;
+      case Actions.ENABLE_NETWORK_INFO:
+        CleverTap.enableDeviceNetworkInfoReporting(true);
+        break;
+      case Actions.ADD_CLEVERTAP_LISTENERS:
+        utils.addCleverTapAPIListeners(true);
+        break;
+      case Actions.REMOVE_CLEVERTAP_LISTENERS:
+        utils.removeCleverTapAPIListeners();
+        break;
+      case Actions.RECORD_EVENT:
+        console.log('name: ', item.name);
+        CleverTap.recordEvent(item.name);
+        break;
+      case Actions.PUSH_PRIMER_HALF_INTERSTITIAL:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == true) {
+            alert('Push Notification permission is already granted');
+          }
+          else {
+            CleverTap.promptPushPrimer({
+              inAppType: 'half-interstitial',
+              titleText: 'Get Notified',
+              messageText:
+                'Please enable notifications on your device to use Push Notifications.',
+              followDeviceOrientation: true,
+              positiveBtnText: 'Allow',
+              negativeBtnText: 'Cancel',
+              backgroundColor: '#FFFFFF',
+              btnBorderColor: '#0000FF',
+              titleTextColor: '#0000FF',
+              messageTextColor: '#000000',
+              btnTextColor: '#FFFFFF',
+              btnBackgroundColor: '#0000FF',
+              btnBorderRadius: '2',
+            });
+          }
+        });
+        break;
+      case Actions.PUSH_PRIMER_HALF_INTERSTITIAL_IMAGE:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == true) {
+            alert('Push Notification permission is already granted');
+          }
+          else {
+            CleverTap.promptPushPrimer({
+              inAppType: 'half-interstitial',
+              titleText: 'Get Notified',
+              messageText:
+                'Please enable notifications on your device to use Push Notifications.',
+              followDeviceOrientation: true,
+              positiveBtnText: 'Allow',
+              negativeBtnText: 'Cancel',
+              backgroundColor: '#FFFFFF',
+              btnBorderColor: '#0000FF',
+              titleTextColor: '#0000FF',
+              messageTextColor: '#000000',
+              btnTextColor: '#FFFFFF',
+              btnBackgroundColor: '#0000FF',
+              imageUrl:
+                'https://clevertap.com/wp-content/themes/clevertap2023/assets/images/ct-logo-2.svg',
+              btnBorderRadius: '2',
+            });
+          }
+        });
+        break;
+      case Actions.PUSH_PRIMER_HALF_INTERSTITIAL_FALLBACK:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == true) {
+            alert('Push Notification permission is already granted');
+          }
+          else {
+            CleverTap.promptPushPrimer({
+              inAppType: 'half-interstitial',
+              titleText: 'Get Notified',
+              messageText:
+                'Please enable notifications on your device to use Push Notifications.',
+              followDeviceOrientation: true,
+              positiveBtnText: 'Allow',
+              negativeBtnText: 'Cancel',
+              backgroundColor: '#FFFFFF',
+              btnBorderColor: '#0000FF',
+              titleTextColor: '#0000FF',
+              messageTextColor: '#000000',
+              btnTextColor: '#FFFFFF',
+              btnBackgroundColor: '#0000FF',
+              btnBorderRadius: '2',
+              fallbackToSettings: true,
+            });
+          }
+        });
+        break;
+      case Actions.PUSH_PRIMER_ALERT:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == true) {
+            alert('Push Notification permission is already granted');
+          }
+          else {
+            CleverTap.promptPushPrimer({
+              inAppType: 'alert',
+              titleText: 'Get Notified',
+              messageText: 'Enable Notification permission',
+              followDeviceOrientation: true,
+              positiveBtnText: 'Allow',
+              negativeBtnText: 'Cancel',
+            });
+          }
+        });
+        break;
+      case Actions.PUSH_PRIMER_ALERT_ORIENTATION:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == true) {
+            alert('Push Notification permission is already granted');
+          }
+          else {
+            CleverTap.promptPushPrimer({
+              inAppType: 'alert',
+              titleText: 'Get Notified',
+              messageText: 'Enable Notification permission',
+              followDeviceOrientation: false,
+              positiveBtnText: 'Allow',
+              negativeBtnText: 'Cancel',
+            });
+          }
+        });
+        break;
+      case Actions.PUSH_PRIMER_ALERT_FALLBACK:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == true) {
+            alert('Push Notification permission is already granted');
+          }
+          else {
+            CleverTap.promptPushPrimer({
+              inAppType: 'alert',
+              titleText: 'Get Notified',
+              messageText: 'Enable Notification permission',
+              followDeviceOrientation: false,
+              positiveBtnText: 'Allow',
+              negativeBtnText: 'Cancel',
+              fallbackToSettings: true,
+            });
+          }
+        });
+        break;
+      case Actions.HARD_PERMISSION:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == false) {
+            CleverTap.promptForPushPermission(false);
+          }
+          else {
+            alert('Push Notification permission is already granted');
+          }
+        });
+        break;
+      case Actions.HARD_PERMISSION_FALLBACK:
+        CleverTap.isPushPermissionGranted((err, res) => {
+          console.log('isPushPermissionGranted', res, err);
+          if (res == false) {
+            CleverTap.promptForPushPermission(true);
+          }
+          else {
+            alert('Push Notification permission is already granted');
+          }
+        });
+        break;
+      case Actions.SYNC_VARIABLES:
+        CleverTap.syncVariables()
+        break;
+      case Actions.GET_VARIABLES:
+        CleverTap.getVariables((err, variables) => {
+          console.log('getVariables: ', variables, err);
+        });
+        break;
+      case Actions.GET_VARIABLE:
+        CleverTap.getVariable('reactnative_var_string', (err, variable) => {
+          console.log(`variable value for key \'reactnative_var_string\': ${variable}`);
+        });
+        break;
+      case Actions.DEFINE_VARIABLES:
+        let variables = {
+          'reactnative_var_string': 'reactnative_var_string_value',
+          'reactnative_var_map': {
+            'reactnative_var_map_string': 'reactnative_var_map_value'
+          },
+          'reactnative_var_int': 6,
+          'reactnative_var_float': 6.9,
+          'reactnative_var_boolean': true
+        };
+        console.log(`Creating variables: ${JSON.stringify(variables)}`);
+        CleverTap.defineVariables(variables);
+        break;
+      case Actions.FETCH_VARIABLES:
+        CleverTap.fetchVariables((err, success) => {
+          console.log('fetchVariables result: ', success);
+        });
+        break;
+      case Actions.VARIABLES_CHANGED:
+        CleverTap.onVariablesChanged((variables) => {
+          console.log('onVariablesChanged: ', variables);
+        });
+        break;
+      case Actions.VALUE_CHANGED:
+        CleverTap.onValueChanged('reactnative_var_string', (variable) => {
+          console.log('onValueChanged: ', variable);
+        });
+        break;
+      case Actions.FETCH_INAPPS:
+        CleverTap.fetchInApps((err, success) => {
+          console.log('fetchInApps result: ', success);
+        });
+        break;
+      case Actions.CLEAR_INAPPS:
+        CleverTap.clearInAppResources(false);
+        break;
+      case Actions.CLEAR_INAPPS_EXPIRED:
+        CleverTap.clearInAppResources(true);
+        break;
+      case Actions.SYNC_CUSTOM_TEMPLATES:
+        CleverTap.syncCustomTemplates();
+        break;
+      case Actions.SYNC_CUSTOM_TEMPLATES_PROD:
+        CleverTap.syncCustomTemplatesInProd(true);
+        break;
+      default:
+        console.warn('Action not recognized:', item.action);
+    }
+  }
+
+  updateLayout = (index) => {
     LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut);
     const array = [...this.state.AccordionData];
     array[index].expanded = !array[index].expanded;
@@ -347,16 +758,17 @@ export default class App extends Component {
 
   render() {
     return (
-      <View style={styles.MainContainer}>
+      <View style={styles.mainContainer}>
         <ScrollView contentContainerStyle={{ paddingHorizontal: 8, paddingVertical: 5 }}>
-          <TouchableOpacity style={styles.button}>
-            <Text style={styles.button_Text}>CleverTap Example</Text>
+          <TouchableOpacity style={styles.header}>
+            <Text style={styles.headerText}>CleverTap Example</Text>
           </TouchableOpacity>
           <CustomTemplate />
           {this.state.AccordionData.map((item, key) => (
-            <Expandable_ListView
+            <ExpandableListView
               key={item.category_Name}
-              onClickFunction={() => this.update_Layout(key)}
+              onToggleView={() => this.updateLayout(key)}
+              onItemPress={this.onItemPress}
               item={item}
             />
           ))}
@@ -367,22 +779,20 @@ export default class App extends Component {
 }
 
 const styles = StyleSheet.create({
-  MainContainer: {
+  mainContainer: {
     flex: 1,
     justifyContent: 'center',
     paddingTop: Platform.OS === 'ios' ? 44 : 0,
     backgroundColor: '#fff',
   },
-  
-  button: {
+  header: {
     backgroundColor: '#fff',
     flexWrap: 'wrap',
     color: '#fff',
     fontSize: 44,
     padding: 10,
   },
-
-  button_Text: {
+  headerText: {
     width: '100%',
     textAlign: 'center',
     color: '#000',
